+++ contracts/contracts/_testing/unit/yield/TestLidoStVaultYieldProvider.sol
@@ -60,8 +60,8 @@ contract TestLidoStVaultYieldProvider is LidoStVaultYieldProvider {
     address _yieldProvider,
     bytes calldata _pubkeys,
     uint64[] calldata _amounts,
-    uint256 _validatorIndex,
-    uint256 _slot,
+    uint64 _validatorIndex,
+    uint64 _slot,
     bytes calldata _withdrawalParamsProof
   ) external view returns (uint256) {
     return _validateUnstakePermissionlessRequest(_yieldProvider, _pubkeys, _amounts, _validatorIndex, _slot, _withdrawalParamsProof);
